Sort- Heap sort algorithm
A menu driven program to sort an array of n integers in ascending order by heap sort algorithm and perform the operations on max heap. Determining time required to sort the elements. Repeat the experiment for different values of n, the number of elements in the array to be sorted and plot a graph of the time taken versus n. The elements can be read from a file or can be generated using the random number generator.